+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed crash when `localStorage` is not available (e.g. due CSP restrictions)
+- Fixed `darkmode` and `darkmodePersistent` settings for `App` to use defaults when coresponding values in options are `undefined`
 
 ## 1.0.0-beta.44 (21-10-2020)